lastitem: optimizations: chdir (unfinished); filter based on readdir info
Unfinished: correct handling of printing (--fullpath or not).

chdir partially so that the OS has less work to do, partially so that
we can store and pass just the filename.

Filtering using the readdir info is avoiding potentially lots of stat
calls. Except, as per docs, in the case of *some* (non-usual, as it
sounds) Unix systems that need to stat to get the info via readdir (in
which case we would do stat *twice*).  (It is confirmed to be much
faster on Linux.)