Create `duu`
Replacement for my Perl script. ~50x faster than it, and more than 2x
faster than `du -s` thanks to parallel execution.

Starting parallel runs before finishing the top-level readdir didn't
give relevant improvements, but leaving it in anyway.

Building a tree as output felt easier to do, and *might* come in
useful for an actual `du` replacement, too.

Error handling is weird maybe: ignores error deeper down, but not in
the toplevel and even first diving (the latter one only reported in
the final loop, though, which is an odd "decision" to be sure).