duu: 'correctly' handle hard-linked files: share their bytes across sites
To know how much of the storage to account for in each site, keep only
a reference to a shared variable entry, and calculate the size share
only in the end.

Oh sh*, perhaps it would have been even more correct to simply divide
bytes by the inode count of the file system, though, no?! This is not
how `du` works, but we don't have to do the same mistakes as it.

Also add some docs.