priorities: parallelize scanning/parsing phase
Get the file list in a single thread (that does not stat the files,
thus is fine), but `par_bridge` to parse each file name (each as a
separate parallel chunk of work!) and do the stat (call
`.metadata()`). Then collect into a Vec because that's the only way to
get it back to the main thread? (Rayon *must* be offering an
alternative here, no? But then need to sort anyway; but then Rayon
could maintain sort order, too.)

In this form, the `sort_by_key()` is needed. Which is why the ids need
to be retained even though they are otherwise not used after calling
`parse_path`.